# storage/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.

package queries

import (
	"database/sql"
	"encoding/json"
	"fmt"
	"time"

	"github.com/google/uuid"
)

type Roles string

const (
	RolesSuperAdmin  Roles = "super_admin"
	RolesBranchAdmin Roles = "branch_admin"
)

func (e *Roles) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = Roles(s)
	case string:
		*e = Roles(s)
	default:
		return fmt.Errorf("unsupported scan type for Roles: %T", src)
	}
	return nil
}

type Attendance struct {
	ID        int32
	Guid      uuid.UUID
	GroupID   sql.NullInt32
	StudentID sql.NullInt32
	IsAttend  sql.NullBool
	Comment   string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type Branch struct {
	ID        int32
	Guid      uuid.UUID
	Address   string
	Phone     string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type Group struct {
	ID        int32
	Guid      uuid.UUID
	Name      string
	SubjectID sql.NullInt32
	Schedule  json.RawMessage
	TeacherID sql.NullInt32
	BranchID  sql.NullInt32
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type GroupStudent struct {
	GroupID   sql.NullInt32
	StudentID sql.NullInt32
}

type Payment struct {
	ID        int32
	Guid      uuid.UUID
	StudentID sql.NullInt32
	GroupID   sql.NullInt32
	Amount    float64
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type Student struct {
	ID        int32
	Guid      uuid.UUID
	Phone     string
	FirstName string
	LastName  string
	BirthDate time.Time
	Address   string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type Subject struct {
	ID        int32
	Guid      uuid.UUID
	Name      string
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type SubjectTeacher struct {
	SubjectID sql.NullInt32
	TeacherID sql.NullInt32
}

type Teacher struct {
	ID        int32
	Guid      uuid.UUID
	Phone     string
	FirstName string
	LastName  string
	BirthDate time.Time
	Address   string
	Salary    float64
	BranchID  sql.NullInt32
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}

type User struct {
	ID        int32
	Guid      uuid.UUID
	Username  string
	Password  string
	Phone     string
	FirstName string
	LastName  string
	IsActive  bool
	BirthDate time.Time
	Address   string
	Role      Roles
	CreatedAt time.Time
	UpdatedAt time.Time
	DeletedAt sql.NullTime
}